requirements.txt
# Core ML and Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Spotify API
spotipy>=2.23.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Dimensionality Reduction
umap-learn>=0.5.3

# Web App
streamlit>=1.25.0

# Model Persistence
joblib>=1.3.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.0.280

# Progress Bars
tqdm>=4.65.0

# Configuration
python-dotenv>=1.0.0
